Move compare methods to their own file
